🚚 Move google-spanner functions to a subdirectory
