+++ src/functions/google-spanner/index.ts
@@ -1,0 +1,2 @@
+export { GoogleSpannerListDatabases } from './list-databases.js';
+export { GoogleSpannerWriteDatabases } from './write-databases.js';
